More whitespace fixes.

+++ src/uk/ac/imperial/lsds/seep/runtimeengine/OutputQueue.java
@@ -4,7 +4,7 @@
  * are made available under the terms of the Eclipse Public License v1.0
  * which accompanies this distribution, and is available at
  * http://www.eclipse.org/legal/epl-v10.html
- * 
+ *
  * Contributors:
  *     Raul Castro Fernandez - initial design and implementation
  ******************************************************************************/
@@ -34,19 +34,19 @@ import com.esotericsoftware.kryo.KryoException;
 import com.esotericsoftware.kryo.io.Output;
 
 public class OutputQueue {
-	
+
 	final private Logger LOG = LoggerFactory.getLogger(OutputQueue.class);
 
 	// replaySemaphore controls whether it is possible to send or not
 	private CoreRE owner = null;
 	private AtomicInteger replaySemaphore = new AtomicInteger(0);
 	private Kryo k = null;
-	
+
 	public OutputQueue(CoreRE owner){
 		this.owner = owner;
 		this.k = initializeKryo();
 	}
-	
+
 	private Kryo initializeKryo(){
 		//optimize here kryo
 		Kryo k = new Kryo();
@@ -57,7 +57,7 @@ public class OutputQueue {
 		k.setAsmEnabled(true);
 		return k;
 	}
-	
+
 	//Start incoming data, one thread has finished replaying
 	public synchronized void start(){
 		/// \todo {this is a safe check that should not be done because we eventually will be sure that it works well}
@@ -74,14 +74,14 @@ public class OutputQueue {
 			this.notify();
 		}
 	}
-	
+
 	public synchronized void stop() {
 		//Stop incoming data, a new thread is replaying
 		LOG.debug("-> replaySemaphore from: {}", replaySemaphore.toString());
 		replaySemaphore.incrementAndGet();
 		LOG.debug("-> replaySemaphore to: {}", replaySemaphore.toString());
 	}
-	
+
 
 	public synchronized void sendToDownstream(DataTuple tuple, EndPoint dest) {
 		SynchronousCommunicationChannel channelRecord = (SynchronousCommunicationChannel) dest;
@@ -159,7 +159,7 @@ public class OutputQueue {
 			ie.printStackTrace();
 		}
 	}
-	
+
 	public void replay(SynchronousCommunicationChannel oi){
 		long a = System.currentTimeMillis();
 				while(oi.getSharedIterator().hasNext()){
@@ -171,12 +171,12 @@ public class OutputQueue {
 		long b = System.currentTimeMillis() - a;
 		System.out.println("Dis.replay: "+b);
 	}
-	
+
 	public void replayTuples(SynchronousCommunicationChannel cci) {
 		Iterator<OutputLogEntry> sharedIterator = cci.getBuffer().iterator();
 		Output output = cci.getOutput();
 		int bufferSize = cci.getBuffer().size();
-		int controlThreshold = (int)(bufferSize)/10;
+		int controlThreshold = (bufferSize)/10;
 		int replayed = 0;
 //		while(sharedIterator.hasNext()) {
 //			BatchTuplePayload dt = sharedIterator.next().batch;
@@ -199,5 +199,5 @@ public class OutputQueue {
 		cci.setSharedIterator(sharedIterator);
 		start();
 	}
-	
+
 }
